soundpath_backend/requirements.txt
django
djangorestframework
django-environ
django-cors-headers
psycopg2-binary
requests
